feat: code changed to modular based files
